[Bronze I] Title: 달팽이는 올라가고 싶다, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/2869. 달팽이는 올라가고 싶다/달팽이는 올라가고 싶다.cc
@@ -3,17 +3,14 @@
 using namespace std;
 
 int main() {
-
-
     int A, B, V;
     cin >> A >> B >> V;
     // 반복문을 쓰면 무조건 시간초과!
     // 무조건 수학적 지식을 활용하여 O(1)로 끝내야함
-    double k = (V - B) / (A - B);
+    int k = (V - B) / (A - B);
     int numer = V - B;
     int denom = A - B;
 
-
-    if (numer % denom == 0) cout << int(k) << endl;
-    else cout << int(k + 1) << endl;
+    if (numer % denom == 0) cout << k << endl;
+    else cout << k + 1 << endl;
 }
